! http: introduce `HttpData` model replacing the byte array in `HttpBody` and `MessageChunk`, closes #365
So far the HTTP model for message entities and chunks was based on "raw" byte arrays thereby violating the immutability "standard".
With this change we introduce a dedicated model for HTTP binary data called `HttpData`, which also adds support for file-based (off-memory) data.

We use the opportunity of this breaking change to refactor the `HttpEntity` model.
`EmptyEntity` and `HttpBody` are gone and replaced with `HttpEntity.Empty` and `HttpEntity.NonEmpty`.
Additionally `MessageChunk#bodyAsString` is gone. Use `chunk.data.asString` instead.